+++ Command/StatusCommand.php
@@ -11,6 +11,7 @@
 
 namespace Massive\Bundle\SearchBundle\Command;
 
+use Massive\Bundle\SearchBundle\Search\SearchManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
 use Symfony\Component\Console\Helper\Table;
 use Symfony\Component\Console\Input\InputInterface;
@@ -22,11 +23,22 @@ use Symfony\Component\Console\Output\OutputInterface;
  */
 class StatusCommand extends ContainerAwareCommand
 {
+    /**
+     * @var SearchManager
+     */
+    private $searchManager;
+
+    public function __construct(SearchManagerInterface $searchManager)
+    {
+        parent::__construct();
+        $this->searchManager = $searchManager;
+    }
+
     public function configure()
     {
         $this->setName('massive:search:status');
         $this->setDescription('Return the status of the configured search engine');
-        $this->setHelp(<<<EOT
+        $this->setHelp(<<<'EOT'
 Return detailed information about the current search implementation
 EOT
         );
@@ -34,8 +46,7 @@ EOT
 
     public function execute(InputInterface $input, OutputInterface $output)
     {
-        $searchManager = $this->getContainer()->get('massive_search.search_manager');
-        $status = $searchManager->getStatus();
+        $status = $this->searchManager->getStatus();
 
         $table = new Table($output);
         $table->setHeaders(['Field', 'Value']);