+++ Tests/Functional/Command/IndexRebuildCommandTest.php
@@ -11,28 +11,58 @@
 
 namespace Massive\Bundle\SearchBundle\Tests\Functional;
 
-use Massive\Bundle\SearchBundle\Command\IndexRebuildCommand;
-use Symfony\Bundle\FrameworkBundle\Console\Application;
-use Symfony\Component\Console\Tester\CommandTester;
-
 class IndexRebuildCommandTest extends BaseTestCase
 {
-    public function setUp()
-    {
-        parent::setUp();
-        $command = new IndexRebuildCommand();
-        $application = new Application($this->getContainer()->get('kernel'));
-        $command->setApplication($application);
-        $this->tester = new CommandTester($command);
-    }
-
     /**
      * This command just fires an event.
      */
     public function testCommand()
     {
-        $this->tester->execute([]);
+        $tester = $this->getCommand('prod', 'massive:search:reindex');
+        $tester->execute([], ['interactive' => false]);
 
-        $this->assertEquals(0, $this->tester->getStatusCode());
+        $this->assertEquals(0, $tester->getStatusCode());
+    }
+
+    /**
+     * It should ask for configuration if a checkpoint exists.
+     */
+    public function testCheckpointResume()
+    {
+        $this->markTestSkipped(
+            'We could do this as detailed here: http://symfony.com/doc/current/components/console/helpers/dialoghelper.html#testing-a-command-which-expects-input'
+        );
+    }
+
+    /**
+     * It show a warning if the environment is not the production environment.
+     */
+    public function testWarningNoEnvNotProd()
+    {
+        $tester = $this->getCommand('dev', 'massive:search:reindex');
+        $tester->execute([], ['interactive' => false]);
+
+        $this->assertContains('WARNING', $tester->getDisplay());
+    }
+
+    /**
+     * It should not show a warning if the environment is the production environment.
+     */
+    public function testNoWarningNoEnvProd()
+    {
+        $tester = $this->getCommand('prod', 'massive:search:reindex');
+        $tester->execute([], ['interactive' => false]);
+
+        $this->assertNotContains('WARNING', $tester->getDisplay());
+    }
+
+    /**
+     * The deprecated command name should still work.
+     */
+    public function testDeprecatedCommand()
+    {
+        $tester = $this->getCommand('prod', 'massive:search:index:rebuild');
+        $tester->execute([], ['interactive' => false]);
+        $this->assertContains('DEPRECATED: The `massive:search:index:rebuild` command is deprecated', $tester->getDisplay());
     }
 }